test add and delete
