+++ package.json
@@ -4,7 +4,8 @@
   "description": "<a name=\"readme-top\"></a>",
   "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
+    "watch": "jest --watch *.js",
     "start": "webpack serve --open",
     "build": "webpack"
   },
@@ -35,9 +36,12 @@
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.11.1",
-    "xml-loader": "^1.2.1"
+    "xml-loader": "^1.2.1",
+    "@babel/plugin-transform-modules-commonjs": "^7.17.9",
+    "jest-environment-jsdom": "^29.3.1"
   },
   "dependencies": {
+    "jest": "^29.3.1",
     "lodash": "^4.17.21"
   }
 }
