+++ src/index.html
@@ -25,7 +25,7 @@
                 <img
                   class="returnImg"
                   id="returnImg"
-                  src="./assets/return.svg"
+                  src="./assets/reload.jpg"
                   alt="returnImg"
                 />
               </button>
