+++ src/style.css
@@ -1,10 +1,7 @@
-* {
-  margin: 0;
-  padding: 0;
-}
-
 body {
   background-color: #dc2e2e;
+  margin: 0;
+  padding: 0;
 }
 
 .container {
